+++ backend/src/util.rs
@@ -2,12 +2,12 @@ use crate::api::GameFetchWarning;
 use crate::game_info_generator::GameInfo;
 use std::collections::HashMap;
 
-pub fn compute_average(times: &[i32]) -> f32 {
-    (times.iter().sum::<i32>() as f32 / times.len() as f32) as f32
+pub fn compute_average(times: &[f32]) -> f32 {
+    (times.iter().sum::<f32>() / times.len() as f32) as f32
 }
 
-pub fn convert_centiseconds_to_seconds(time: f32) -> f32 {
-    (time / 100 as f32) as f32
+pub fn convert_centiseconds_to_seconds(time: i32) -> f32 {
+    (time as f32 / 100.0) as f32
 }
 
 pub fn generate_dummy_erros_testing(skipped_games: &mut HashMap<usize, GameFetchWarning>) {
